+++ src/marshmallow/fields.py
@@ -164,6 +164,9 @@ class Field(typing.Generic[_InternalType]):
 
     .. versionchanged:: 3.13.0
         Replace ``missing`` and ``default`` parameters with ``load_default`` and ``dump_default``.
+
+    .. versionchanged:: 4.0.0
+        Remove ``context`` property.
     """
 
     # Some fields, such as Method fields and Function fields, are not expected
@@ -442,15 +445,6 @@ class Field(typing.Generic[_InternalType]):
         """
         return value
 
-    # Properties
-
-    @property
-    def context(self) -> dict | None:
-        """The context dictionary for the parent :class:`Schema`."""
-        if self.parent:
-            return self.parent.context
-        return None
-
 
 class Raw(Field[typing.Any]):
     """Field that applies no formatting."""
@@ -540,8 +534,6 @@ class Nested(Field):
     def schema(self) -> Schema:
         """The nested Schema object."""
         if not self._schema:
-            # Inherit context from parent.
-            context = getattr(self.parent, "context", {})
             if callable(self.nested) and not isinstance(self.nested, type):
                 nested = self.nested()
             else:
@@ -554,7 +546,6 @@ class Nested(Field):
 
             if isinstance(nested, Schema):
                 self._schema = copy.copy(nested)
-                self._schema.context.update(context)
                 # Respect only and exclude passed from parent and re-initialize fields
                 set_class = typing.cast(type[set], self._schema.set_class)
                 if self.only is not None:
@@ -581,7 +572,6 @@ class Nested(Field):
                     many=self.many,
                     only=self.only,
                     exclude=self.exclude,
-                    context=context,
                     load_only=self._nested_normalized_option("load_only"),
                     dump_only=self._nested_normalized_option("dump_only"),
                 )
@@ -1994,14 +1984,12 @@ class Function(Field):
 
     :param serialize: A callable from which to retrieve the value.
         The function must take a single argument ``obj`` which is the object
-        to be serialized. It can also optionally take a ``context`` argument,
-        which is a dictionary of context variables passed to the serializer.
+        to be serialized.
         If no callable is provided then the ```load_only``` flag will be set
         to True.
     :param deserialize: A callable from which to retrieve the value.
         The function must take a single argument ``value`` which is the value
-        to be deserialized. It can also optionally take a ``context`` argument,
-        which is a dictionary of context variables passed to the deserializer.
+        to be deserialized.
         If no callable is provided then ```value``` will be passed through
         unchanged.
 
@@ -2010,6 +1998,9 @@ class Function(Field):
 
     .. versionchanged:: 3.0.0a1
         Removed ``func`` parameter.
+
+    .. versionchanged:: 4.0.0
+        Don't pass context to serialization and deserialization functions.
     """
 
     _CHECK_ATTRIBUTE = False
@@ -2036,20 +2027,12 @@ class Function(Field):
         self.deserialize_func = deserialize and utils.callable_or_raise(deserialize)
 
     def _serialize(self, value, attr, obj, **kwargs):
-        return self._call_or_raise(self.serialize_func, obj, attr)
+        return self.serialize_func(obj)
 
     def _deserialize(self, value, attr, data, **kwargs):
         if self.deserialize_func:
-            return self._call_or_raise(self.deserialize_func, value, attr)
+            return self.deserialize_func(value)
         return value
-
-    def _call_or_raise(self, func, value, attr):
-        if len(utils.get_func_args(func)) > 1:
-            if self.parent.context is None:
-                msg = f"No context available for Function field {attr!r}"
-                raise ValidationError(msg)
-            return func(value, self.parent.context)
-        return func(value)
 
 
 _ContantType = typing.TypeVar("_ContantType")